Change handlers structure
